mean rework: returns null for no values regardless of the type. Added orNull overloads for each mean function. Added specific overloads for each primitive type -> Double(?) and big number -> BigDecimal(?)

+++ core/src/test/kotlin/org/jetbrains/kotlinx/dataframe/puzzles/MediumTests.kt
@@ -66,7 +66,7 @@ class MediumTests {
             -1, 0, 1,
         )
 
-        df.convert { colsOf<Double>() }.with { (it - rowMean()).roundToInt() } shouldBe expected
+        df.convert { colsOf<Double>() }.with { (it - rowMean().toDouble()).roundToInt() } shouldBe expected
     }
 
     @Test
